refactor(entry): buildTagDashboard を entry/domain へ移動
tag-dashboard.ts は DB / supabase 依存ゼロの pure transformation で、

Calendar / Tags / Chronotype 既存 domain と同じ責務形。entry/server は

DB 取得 + tRPC orchestration に専念する形に近づける。

- git mv で 2 ファイル移動 (本体 + test)

  - entry/server/tag-dashboard.ts → entry/domain/tag-dashboard.ts

  - entry/server/__tests__/tag-dashboard.test.ts → entry/domain/__tests__/

- entry/server/tag-statistics.ts の import path 更新

- entry/domain/index.ts barrel に buildTagDashboard と関連型を追加

feature 公開 barrel の export 範囲は不変。entry 内部消費のみ。
